Fix most warnings in openj9.dtfj
* don't use raw types
* add missing @Override annotations
* remove redundant casts
* properly reference static features
* remove misplaced javadoc
* fix and simplify Symbol.SymbolComparator
* simplify comparator in MemoryRange
* use modern API, e.g. Collections.emptyIterator()
* suppress some unavoidable serialization warnings
* remove unused, empty collections
* fix loss of precision due to implicit casts
* remove suppression of rawtypes and unchecked warnings
* remove dead code, unused fields
* fix infinite recursion in SegmentSortException.getCause()
* add explicit constructors
* avoid using deprecated API
* remove obsolete API related to "packed"
* prefer interfaces over concrete types (List instead of ArrayList)
* mark classes final to avoid most "this-escape" warnings
* make GCIterator & SlotIterator classes generic

+++ debugtools/DDR_VM/src/com/ibm/j9ddr/vm29/tools/ddrinteractive/commands/WalkInternTableCommand.java
@@ -297,7 +297,7 @@ public class WalkInternTableCommand extends Command
 	 * @return void
 	 */
 	private void walkSharedTable(PrintStream out) throws CorruptDataException {
-		int totalWeight = 0;
+		long totalWeight = 0;
 		J9JavaVMPointer vm = J9RASHelper.getVM(DataType.getJ9RASPointer());
 		J9SharedInvariantInternTablePointer sharedInternTable = vm.sharedInvariantInternTable();
 		/* Check if there is sharedInternTable before to go further */
@@ -345,7 +345,7 @@ public class WalkInternTableCommand extends Command
 	 * 
 	 */
 	private void walkLocalTable(PrintStream out) throws CorruptDataException {
-		int totalWeight = 0;
+		long totalWeight = 0;
 		J9DbgStringInternTablePointer stringInternTablePtr = getRomClassBuilderPtr(out);
 		
 		if (stringInternTablePtr.isNull()) {
